Move test to primitives
